requirements.txt
alembic==0.9.1
appdirs==1.4.3
click==6.7
coverage==4.4.1
dominate==2.3.1
flake8==3.3.0
Flask==0.12.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.2
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-Testing==0.6.2
Flask-WTF==0.14.2
itsdangerous==0.24
MarkupSafe==1.0
marshmallow==2.13.5
mccabe==0.6.1
nose==1.3.7
packaging==16.8
psycopg2==2.7.1
pycodestyle==2.3.1
pyflakes==1.5.0
PyJWT==1.5.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2017.2
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.7
visitor==0.1.3
Werkzeug==0.12.1
